+++ inst/examples/AcuDNPercentile_ex.R
@@ -6,5 +6,5 @@ AcuDNPercentile(df = exampleHourlyData,
                 parameter = "leq",
                 from = "5",
                 to = "22",
-                period = "night")
+                period = "night")[1:5]