+++ inst/examples/RoundTo_ex.R
@@ -1,3 +1,3 @@
 #data("P1FA")
 
-RoundTo(x = P1FA$LAeq, multiple = 0.5)
+RoundTo(x = P1FA$LAeq, multiple = 0.5)[1:10]