+++ inst/examples/runningLeq_ex.R
@@ -1,3 +1,3 @@
 #data("P1FA")
 
-runningLeq(x = P1FA$LAeq)
+runningLeq(x = P1FA$LAeq)[1:10]